+++ CHANGELOG.md
@@ -33,6 +33,8 @@ This is just a re-release of 10.4.0 due to an issue with the 10.4.0 release.
     - #1667
     - #1813
 
+- Fix completions for alias `fdfind` in deb release, see #1888 (@skane-lukas)
+
 ## Changes
 
 - Minimum required rust version has been increased to 1.90.0. Notably, this means dropping fully support for intel Mac and Windows 7.